[#21] Restore timezone when reading work history from storage
With the current approach, for simplicity, timezone is not stored
in the TSV. Some piece of information is lost. This change introduces
restoring this piece of information when reading the history. It's crucial
to have the right timezone, to be able to calculate e.g. lenght of the
current work item.

The current approach assumes that the user is constantly in the same timezone.